count flip bits

+++ bit_manipulation/5-flip_bits.c
@@ -7,7 +7,7 @@
  **/
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int diff = n ^ m;
+	unsigned long int diff = n ^ m;
 	unsigned int count = 0;
 
 	while (diff != 0)
